refactor: update event includes to use net/event namespace and adjust file paths
