+++ sese/internal/win/net/event/WSAEventConvert.cpp
@@ -12,18 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/**
- * @file KqueueEvent.h
- * @brief kqueue event
- * @author kaoru
- */
+#include "sese/internal/win/net/event/WSAEventConvert.h"
 
-#pragma once
+#include <winsock2.h>
 
-#include "sese/event/BaseEvent.h"
+unsigned int sese::event::WSAEventConvert::fromNativeEvent(int event) {
+    unsigned result = 0;
+    if (event & FD_READ) {
+        result |= EVENT_READ;
+    }
+    if (event & FD_WRITE) {
+        result |= EVENT_WRITE;
+    }
+    return result;
+}
 
-namespace sese::event {
-/// kqueue event
-class KqueueEvent : public BaseEvent {
-};
-} // namespace sese::event
+int sese::event::WSAEventConvert::toNativeEvent(unsigned int event) {
+    int result = 0;
+    if (event & EVENT_READ) {
+        result |= FD_READ;
+    }
+    if (event & EVENT_WRITE) {
+        result |= FD_WRITE;
+    }
+    return result;
+}
