+++ sese/internal/win/net/event/WSAEventLoop.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "sese/native/win/event/WSAEventLoop.h"
+#include "sese/internal/win/net/event/WSAEventLoop.h"
 
 #include <winsock2.h>
