+++ sese/internal/win/system/FileLocker.cpp
