+++ sese/internal/win/system/NetworkUtil.cpp
