+++ sese/internal/win/system/Semaphore.cpp
