+++ sese/internal/win/system/SharedMemory.cpp
