+++ sese/internal/win/thread/GlobalThreadPool.cpp
