+++ sese/net/event/BaseEvent.h
