+++ sese/service/iocp/IOCPServer.h
@@ -23,7 +23,7 @@
 #include <sese/Config.h>
 
 #if defined(SESE_PLATFORM_WINDOWS)
-#include <sese/native/win/service/iocp/NativeIOCPServer_V1.h>
+#include <sese/internal/win/service/iocp/NativeIOCPServer_V1.h>
 #else
 #include <sese/service/iocp/IOCPServer_V1.h>
 #endif
